+++ app/docker/entrypoint.sh
@@ -13,7 +13,7 @@ sed -i "s|\$DB_NAME|$DB_NAME|g" $env_file
 sed -i "s|\$DB_USER|$DB_USER|g" $env_file
 sed -i "s|\$DB_PASSWORD|$DB_PASSWORD|g" $env_file
 sed -i "s|\$DB_HOST|$DB_HOST|g" $env_file
-sed -i "s|\$DOMAIN|${APP_DOMAIN:-$DOMAIN}|g" $env_file
+sed -i "s|\$APP_URL|$APP_URL|g" $env_file
 sed -i "s|\$SMTP_FROM|$SMTP_FROM|g" $env_file
 sed -i "s|\$ENABLE_AZURE_OAUTH|$ENABLE_AZURE_OAUTH|g" $env_file
 sed -i "s|\$OAUTH_AZURE_CLIENT_ID|$OAUTH_AZURE_CLIENT_ID|g" $env_file