+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Migration notes
 
+- You can remove the `APP_DOMAIN` variable from the root `.env` file and you must add a `APP_URL` variable (ex. `https://agentj.example.com`).
 - You must add a `SMTP_FROM` variable in the root `.env` file as it's no longer computed from the domain. Recommended value is `no-reply@<domain>`.
 
 ## 2025-08-25 - 2.2.1